// frontend/src/shared/api/__generated__/schema.ts
export const userCreateRequestBodyDtoSchema = z.object({
  name: z.string(),
  grade: z.number(),
  memo: z.string().optional().nullable(),
});